@@ -1,12 +1,12 @@
 #!/usr/bin/env python3
 # -*- Mode: python; tab-width: 4; indent-tabs-mode:nil; coding:utf-8 -*-
 #
-# Copyright (c) 2020 Authors and contributors
+# Copyright (c) 2022 Authors and contributors
 # (see the file AUTHORS for the full list of names)
 #
 # Released under the GNU Public Licence, v2 or any higher version
 # SPDX-License-Identifier: GPL-2.0-or-later
-
+import logging
 import os
 import subprocess
 import tempfile
@@ -14,12 +14,17 @@ import tempfile
 import numpy as np
 import MDAnalysis as mda
 
-from .base import SingleGroupAnalysisBase
+from .base import AnalysisBase, PlanarBase
 from .. import tables
+from ..decorators import (
+    make_whole,
+    set_planar_class_doc,
+    set_verbose_doc,
+)
 from ..lib import sfactor
 from ..utils import check_compound, savetxt
-from ..decorators import planar_base
 
+logger = logging.getLogger(__name__)
 
 def compute_form_factor(q, atom_type):
     """Calculate the form factor for the given element for given q (1/nm).
@@ -57,8 +62,8 @@ def compute_form_factor(q, atom_type):
 
     return form_factor
 
-
-class saxs(SingleGroupAnalysisBase):
+@set_verbose_doc
+class Saxs(AnalysisBase):
     """Compute SAXS scattering intensities.
 
     The q vectors are binned
@@ -71,57 +76,59 @@ class saxs(SingleGroupAnalysisBase):
     based on Cromer-Mann parameters. By using the -sel option atoms can be selected for which the
     profile is calculated. The selection uses the MDAnalysis selection commands.
 
-    **Inputs**
+    Parameters
+    ----------
+    atomgroup : AtomGroup
+       Atomgroup on which the analysis is executed
+    noboindata : bool
+        Do not bin the data. Only works reliable for NVT!
+    startq : float
+        Starting q (1/nm)
+    endq : float
+        Ending q (1/nm)
+    dq : float
+        binwidth (1/nm)
+    mintheta float
+        Minimal angle (°) between the q vectors and the z-axis.
+    maxtheta : float
+        Maximal angle (°) between the q vectors and the z-axis.
+    output : str
+        Output filename
+    concfreq : int
+        Default number of frames after which results are calculated and files refreshed.
+        If `0` results are only calculated at the end of the analysis and not
+        saved by default.
+    ${VERBOSE_PARAMETER}
 
-    :param outfreq (float): Number of frames after which the output is updated.
-    :param output (str): Output filename
-    :param noboindata (bool): Do not bin the data. Only works reliable for NVT!
-    :param startq (float): Starting q (1/nm)
-    :param endq (float): Ending q (1/nm)
-    :param dq (float): binwidth (1/nm)
-    :param mintheta (float): Minimal angle (°) between the q vectors and the z-axis.
-    :param maxtheta (float): Maximal angle (°) between the q vectors and the z-axis.
-
-
-    **Attributes**
-
-    :returns results.q (numpy.ndarray): length of binned q-vectors
-    :returns results.q_indices (numpy.ndarray): Miller indices of q-vector (only if noboindata==True)
-    :returns results.scat_factor(numpy.ndarray): Scattering intensities
-
-
+    Attributes
+    ----------
+    results.q : np.ndarray
+        length of binned q-vectors
+    results.q_indices : np.ndarray
+        Miller indices of q-vector (only if noboindata==True)
+    results.scat_factor : np.ndarray
+        Scattering intensities
     """
-
     def __init__(self,
                  atomgroup,
-                 outfreq=100,
-                 output="sq.dat",
                  nobin=False,
                  startq=0,
                  endq=60,
                  dq=0.05,
                  mintheta=0,
                  maxtheta=180,
+                 output="sq.dat",
+                 concfreq=0,
                  **kwargs):
-        super().__init__(atomgroup, **kwargs)
-        self.outfreq = outfreq
-        self.output = output
+        super(Saxs, self).__init__(atomgroup, **kwargs)
         self.nobindata = nobin
         self.startq = startq
         self.endq = endq
         self.dq = dq
         self.mintheta = mintheta
         self.maxtheta = maxtheta
-
-    def _configure_parser(self, parser):
-        parser.add_argument('-dout', dest='outfreq')
-        parser.add_argument('-sq', dest='output')
-        parser.add_argument('-startq', dest='startq')
-        parser.add_argument('-endq', dest='endq')
-        parser.add_argument('-nobin', dest='nobindata')
-        parser.add_argument('-dq', dest='dq', type=float, default=0.05, help='')
-        parser.add_argument('-mintheta', dest='mintheta')
-        parser.add_argument('-maxtheta', dest='maxtheta')
+        self.output = output
+        self.concfreq = concfreq
 
     def _prepare(self):
 
@@ -142,26 +149,22 @@ class saxs(SingleGroupAnalysisBase):
 
         self.groups = []
         self.atom_types = []
-        if self._verbose:
-            print("\nMap the following atomtypes:")
+        logger.info("\nMap the following atomtypes:")
         for atom_type in np.unique(self.atomgroup.types).astype(str):
             try:
                 element = tables.atomtypes[atom_type]
             except KeyError:
-                raise RuntimeError(
-                    "No suitable element for '{0}' found. You can add '{0}' together with a suitable element to 'share/atomtypes.dat'."
-                    .format(atom_type))
+                raise RuntimeError(f"No suitable element for '{atom_type}' "
+                                   f"found. You can add '{atom_type}' "
+                                   "together with a suitable element "
+                                   "to 'share/atomtypes.dat'.")
             if element == "DUM":
                 continue
             self.groups.append(
                 self.atomgroup.select_atoms("type {}*".format(atom_type)))
             self.atom_types.append(atom_type)
 
-            if self._verbose:
-                print("{:>14} --> {:>5}".format(atom_type, element))
-
-        if self._verbose:
-            print("")
+            logger.info("{:>14} --> {:>5}".format(atom_type, element))
 
         if self.nobindata:
             self.box = np.diag(
@@ -207,11 +210,12 @@ class saxs(SingleGroupAnalysisBase):
                                               bins=np.arange(self.nbins + 1))[0]
                 self.struct_factor[:, i] += np.nan_to_num(struct_ts)
 
-        if self._save and self._frame_index % self.outfreq == 0 and self._frame_index > 0:
-            self._calculate_results()
-            self._save_results()
+        if self.concfreq and self._frame_index % self.concfreq == 0 \
+            and self._frame_index > 0:
+            self._conclude()
+            self.save()
 
-    def _calculate_results(self):
+    def _conclude(self):
         self._index = self._frame_index + 1
         if self.nobindata:
             self.results.scat_factor = self.S_array.sum(axis=3)
@@ -230,7 +234,7 @@ class saxs(SingleGroupAnalysisBase):
 
         self.results.scat_factor /= (self._index * self.atomgroup.n_atoms)
 
-    def _save_results(self):
+    def save(self):
         """Saves the current profiles to a file."""
 
         if self.nobindata:
@@ -258,44 +262,55 @@ class saxs(SingleGroupAnalysisBase):
                     fmt='%.4e')
 
 
-class debye(SingleGroupAnalysisBase):
+@set_verbose_doc
+class Debye(AnalysisBase):
     """Calculate scattering intensities using the debye equation.
 
-        **Inputs**
+    Parameters
+    ----------
+    atomgroup : AtomGroup
+       Atomgroup on which the analysis is executed
+    outfreq float :
+        Number of frames after which the output is updated.
+    output : str
+        Output filename
+    startq : float
+        Starting q (1/nm)
+    endq : float
+        Ending q (1/nm)
+    dq : float
+        binwidth (1/nm)
+    sinc : bool
+        Apply sinc damping
+    debyer : str
+        Path to the debyer executable
+    ${VERBOSE_PARAMETER}
 
-       :param outfreq (float): Number of frames after which the output is updated.
-       :param output (str): Output filename
-       :param startq (float): Starting q (1/nm)
-       :param endq (float): Ending q (1/nm)
-       :param dq (float): binwidth (1/nm)
-       :param sinc (bool): Apply sinc damping
-       :param debyer (str): Path to the debyer executable
-
-        **Attributes**
-
-        :returns results.q (numpy.ndarray): length of binned q-vectors
-        :returns results.scat_factor (numpy.ndarray): Scattering intensities
+    Attributes
+    ----------
+    results.q : np.ndarray
+        length of binned q-vectors
+    results.scat_factor : np.ndarray
+        Scattering intensities
     """
-
     def __init__(self,
                  atomgroup,
-                 sel="all",
-                 outfreq=100,
-                 output="sq.dat",
                  startq=0,
                  endq=60,
                  dq=0.05,
                  sinc=False,
                  debyer="debyer",
+                 output="sq.dat",
+                 concfreq=0,
                  **kwargs):
-        super().__init__(atomgroup, **kwargs)
-        self.outfreq = outfreq
-        self.output = output
+        super(Debye, self).__init__(atomgroup, **kwargs)
         self.startq = startq
         self.endq = endq
         self.dq = dq
         self.sinc = sinc
         self.debyer = debyer
+        self.output = output
+        self.concfreq = concfreq
 
     def _configure_parser(self, parser):
         parser.add_argument('-dout', dest='outfreq')
@@ -389,11 +404,12 @@ class debye(SingleGroupAnalysisBase):
                         stderr=self._OUT,
                         shell=True)
 
-        if self._save and self._frame_index % self.outfreq == 0 and self._frame_index > 0:
-            self._calculate_results()
-            self._save_results()
+        if self.concfreq and self._frame_index % self.concfreq == 0 \
+            and self._frame_index > 0:
+            self._conclude()
+            self.save()
 
-    def _calculate_results(self):
+    def _conclude(self):
         datfiles = [f for f in os.listdir(self._tmp) if f.endswith(".dat")]
 
         s_tmp = np.loadtxt("{}/{}".format(self._tmp, datfiles[0]))
@@ -414,81 +430,89 @@ class debye(SingleGroupAnalysisBase):
         self.results.q = 10 * q[nonzeros]
         self.results.scat_factor = s_out[nonzeros] / len(datfiles)
 
-    def _conclude(self):
-        for f in os.listdir(self._tmp):
-            os.remove("{}/{}".format(self._tmp, f))
+        # Remove temp files at the end of the analysis
+        if self.n_frames == self._frame_index + 1:
+            for f in os.listdir(self._tmp):
+                os.remove("{}/{}".format(self._tmp, f))
 
-        os.rmdir(self._tmp)
+            os.rmdir(self._tmp)
 
-    def _save_results(self):
+    def save(self):
         savetxt(self.output,
                 np.vstack([self.results.q, self.results.scat_factor]).T,
                 header="q (1/A)\tS(q)_tot (arb. units)",
                 fmt='%.8e')
 
 
-@planar_base()
-class diporder(SingleGroupAnalysisBase):
+@set_verbose_doc
+@set_planar_class_doc
+@make_whole()
+class Diporder(PlanarBase):
     """Calculation of dipolar order parameters.
-    
-    Calculations include the projected dipole density 
-    P_0⋅ρ(z)⋅cos(θ[z]), the dipole orientation cos(θ[z]), the squared dipole 
+
+    Calculations include the projected dipole density
+    P_0⋅ρ(z)⋅cos(θ[z]), the dipole orientation cos(θ[z]), the squared dipole
     orientation cos²(Θ[z]) and the number density ρ(z).
 
-    **Inputs**
+    Parameters
+    ----------
+    atomgroup : AtomGroup
+       Atomgroup on which the analysis is executed
+    ${PLANAR_CLASS_PARAMETERS}
+    sym : bool
+        symmetrize the profiles
+    binmethod : str
+        binning method: center of mass (COM), center of charge (COC)
+        or oxygen position (OXY). Note: `OXY` only works for water
+        oxygens with name `OW*`
+    ${MAKE_WHOLE_PARAMETER}
+    output : str
+        Output filename
+    concfreq : int
+        Default number of frames after which results are calculated and files refreshed.
+        If `0` results are only calculated at the end of the analysis and not
+        saved by default.
+    ${VERBOSE_PARAMETER}
 
-    :param output (str): Output filename
-    :param outfreq (int): Default number of frames after which output files are
-                         refreshed
-    :param bsym (bool): symmetrize the profiles
-    :param binmethod (str): binning method: center of Mass (COM), center of charge (COC)
-                            or oxygen position (OXY).
-                            Note: `OXY` only works for water oxygens with name `OW*`
-    :param bpbc (bool): do not make broken molecules whole again
-                       (only works if molecule is smaller than shortest box vector
-
-    
-    **Attributes**
-
-    :returns results.z (numpy.ndarray): bins [nm]
-    :returns results.P0 (numpy.ndarray): P_0⋅ρ(z)⋅cos(θ[z]) [e/nm²]
-    :returns results.cos_theta (numpy.ndarray): cos(θ[z])
-    :returns results.cos_2_theta (numpy.ndarray): cos²(Θ[z])
-    :returns results.rho (numpy.ndarray): ρ(z) [1/nm³]
-
+    Attributes
+    ----------
+    ${PLANAR_CLASS_ATTRIBUTES}
+    results.P0 : np.ndarray
+        P_0⋅ρ(z)⋅cos(θ[z]) [e/nm²]
+    results.cos_theta : np.ndarray
+        cos(θ[z])
+    results.cos_2_theta : np.ndarray
+        cos²(Θ[z])
+    results.rho : np.ndarray
+        ρ(z) [1/nm³]
     """
-
     def __init__(self,
                  atomgroup,
-                 output="diporder.dat",
-                 outfreq=10000,
-                 bsym=False,
-                 binmethod='COM',
-                 bpbc=True,
-                 # Planar base arguments are necessary for buidling CLI
                  dim=2,
                  binwidth=0.1,
                  center=False,
                  comgroup=None,
+                 sym=False,
+                 binmethod='COM',
+                 make_whole=True,
+                 output="diporder.dat",
+                 concfreq=0,
                  **kwargs):
-        super().__init__(atomgroup, **kwargs)
-
-        self.output = output
-        self.outfreq = outfreq
-        self.bsym = bsym
+        super(Diporder, self).__init__(atomgroups=atomgroup,
+                                       dim=dim,
+                                       binwidth=binwidth,
+                                       center=center,
+                                       comgroup=comgroup,
+                                       **kwargs)
+        self.sym = sym
         self.binmethod = binmethod
-        self.bpbc = bpbc
-
-    def _configure_parser(self, parser):
-        parser.add_argument('-o', dest='output')
-        parser.add_argument('-dout', dest="outfreq")
-        parser.add_argument('-sym', dest='bsym')
-        parser.add_argument('-bin', dest='binmethod')
-        parser.add_argument('-nopbcrepair', dest='bpbc')
+        self.make_whole = make_whole
+        self.concfreq = concfreq
+        self.output = output
 
     def _prepare(self):
         """Set things up before the analysis loop begins"""
-
+        super(Diporder, self)._prepare()
         self.binmethod = self.binmethod.upper()
         if self.binmethod not in ["COM", "COC", "OXY"]:
             raise ValueError('Unknown binning method: {}'.format(
@@ -518,10 +542,7 @@ class diporder(SingleGroupAnalysisBase):
         self.unit[self.dim] += 1
 
     def _single_frame(self):
-        # Make molecules whole
-        if self.bpbc:
-            self.atomgroup.unwrap(compound=check_compound(self.atomgroup))
-
+        super(Diporder, self)._single_frame()
         dz_frame = self._ts.dimensions[self.dim] / self.n_bins
 
         chargepos = self.atomgroup.positions * self.atomgroup.charges[:, np.
@@ -562,15 +583,17 @@ class diporder(SingleGroupAnalysisBase):
                 self.unit)**2)[0]
         self.rho += bincount / (A * dz_frame / 1e3)  # convert to 1/nm^3
 
-        if self._save and self._frame_index % self.outfreq == 0 and self._frame_index > 0:
-            self._calculate_results()
-            self._save_results()
+        if self.concfreq and self._frame_index % self.concfreq == 0 \
+            and self._frame_index > 0:
+            self._conclude()
+            self.save()
 
-    def _calculate_results(self):
+    def _conclude(self):
         """Calculate the results.
 
         Called at the end of the run() method to before the _conclude function.
         Can also called during a run to update the results during processing."""
+        super(Diporder, self)._conclude()
         self._index = self._frame_index + 1
         self.results.P0 = self.P0 / self._frame_index
         self.results.rho= self.rho / self._frame_index
@@ -579,7 +602,7 @@ class diporder(SingleGroupAnalysisBase):
             self.results.cos_theta = np.nan_to_num(self.cos_theta / self.bin_count)
             self.results.cos_2_theta = np.nan_to_num(self.cos_2_theta / self.bin_count)
 
-        if self.bsym:
+        if self.sym:
             for i in range(len(self.results.z) - 1):
                 self.results.z[i +
                                   1] = .5 * (self.results.z[i + 1] +
@@ -588,11 +611,8 @@ class diporder(SingleGroupAnalysisBase):
                     i + 1] = .5 * (self.results.diporder[i + 1] +
                                    self.results.diporder[i + 1][-1::-1])
 
-    def _save_results(self):
-        """Save results to a file.
-
-        Called at the end of the run() method after _calculate_results and
-        _conclude"""
+    def save(self):
+        """Save results to a file."""
 
         header = "z [nm]\t"
         header += "P_0*rho(z)*cos(Theta[z]) [e/nm^2]\t"
