@@ -12,207 +12,55 @@ import warnings
 
 import numpy as np
 
-from .arg_completion import _append_to_doc
+from .utils import check_compound
 
 
-def planar_base():
-    """Class Decorator to provide options and attributes for analysis in planar
-    cofinenement. Decorator will add the options `dim`, `bindwidth`, 
-    `comgroup`, `center` to the AnalysisClass.
-    `dim` controls the direction of binning and `bindwidth` the 
-    bindwidth in nanometers. With `comgroup` the binning is performed relative 
-    to the center of mass of the selected group. With `center` the 
-    binning is preformed relative to the center of the (changing) box.
+verbose_parameter_doc = (
+    """verbose : bool
+        Turn on more logging and debugging"""
+)
 
-    Furthermore the item "z" is added automatically to the classes 
-    results dictionary.
-    """
-    def inner(original_class):
-        # Make copy of original functions, so we can call them
-        # without recursion problems.
-        orig_init = original_class.__init__
-        orig_configure_parser = original_class._configure_parser
-        orig_prepare = original_class._prepare
-        orig_single_frame = original_class._single_frame
-        orig_calculate_results = original_class._calculate_results
+planar_class_parameters_doc = (
+    """dim : int
+        Dimension for binning (x=0, y=1, z=2)
+    binwidth : float
+        binwidth (nanometer)
+    comgroup : AtomGroup
+        Perform the binning relative to the center of mass of the
+        selected group.
+    center : bool
+        Perform the binning relative to the center of the (changing) box."""
+)
 
-        @functools.wraps(original_class.__init__)
-        def __init__(self,
-                     atomgroups,
-                     *args,
-                     dim=2,
-                     binwidth=0.1,
-                     center=False,
-                     comgroup=None,
-                     **kwargs):
-            # Call the original __init__ first to append arguments.
-            orig_init(self, atomgroups, *args, **kwargs)
-            self.dim = dim
-            self.binwidth = binwidth
-            self.center = center
-            self.comgroup = comgroup
+planar_class_attributes_doc = (
+    """results.z : list
+        bins"""
+)
 
-            # TODO: Try to manipulate signature to get rid of standard 
-            # arguments in decorated planar classes.
-            # Or find a way to extract default arguments, while building CLI...
+make_whole_parameter_doc = (
+    """make_whole : bool
+        Make molecules whole; If the input already contains whole molecules
+        this can be disabled to gain speedup"""
+)
 
-        def _configure_parser(self, parser):
-            # Call the original _configure_parser first to append arguments.
-            orig_configure_parser(self, parser)
-            parser.add_argument('-d', dest='dim')
-            parser.add_argument('-dz', dest='binwidth')
-            parser.add_argument('-com', dest='comgroup')
-            parser.add_argument('-center', dest='center')
 
-        def _prepare(self):
-            if self.dim not in [0, 1, 2]:
-                raise ValueError("Dimension can only be 0=X or 1=Y or 2=Z.")
+def set_verbose_doc(public_api):
+    if public_api.__doc__ is not None:
+        public_api.__doc__ = public_api.__doc__.replace(
+            "${VERBOSE_PARAMETER}",
+            verbose_parameter_doc)
+    return public_api
 
-            # Workaround since currently not alle module have option 
-            # with zmax and zmin
-            if not hasattr(self, '_zmax'):
-                self._zmax = None
 
-            if self._zmax is None:
-                self.Lz = 0
-                self.zmax = self._universe.dimensions[self.dim]
-            else:
-                self.zmax = 10 * self._zmax
-
-            if not hasattr(self, 'zmin'):
-                self.zmin = 0
-            self.zmin *= 10
-            self.binwidth *= 10
-
-            self.n_bins = int(np.ceil((self.zmax - self.zmin) / self.binwidth))
-
-            if self._verbose:
-                print('Using', self.n_bins, 'bins.')
-
-            if self.comgroup is not None:
-                self.comsel = self._universe.select_atoms(self.comgroup)
-                if self._verbose:
-                    print("Performing the binning relative to the "
-                          "center of mass of '{}' containing {} atoms.".format(
-                          self.comgroup, self.comsel.n_atoms))
-                if self.comsel.n_atoms == 0:
-                    raise ValueError(
-                        "`{}` does not contain any atoms. Please adjust "
-                        "'com' selection.".format(self.comgroup))
-            if self.comgroup is not None:
-                self.center = True  # always center when COM
-
-            orig_prepare(self)
-
-        def get_bins(self, positions, dim=None):
-            """"Calculates bins based on given positions. dim denotes 
-            the dimension for calculating bins. If `None` the default 
-            dim is taken.
-            
-            Attributes
-            ----------
-            positions : numpy.ndarray
-                3 dimensional positions
-            dim : int
-                dimesion for binning (x=0, y=1, z=2)"""
-            dim = self.dim if dim is None else dim
-            dz = self._ts.dimensions[dim] / self.n_bins
-            bins = np.rint(positions[:, dim] / dz) 
-            bins %= self.n_bins
-            return bins.astype(int)
-
-        def _single_frame(self, *args, **kwargs):
-
-            if self._zmax is None:
-                self.zmax = self._ts.dimensions[self.dim]
-                self.Lz += self.zmax
-                
-
-            # Center of mass calculation with generalization to periodic systems
-            # see Bai, Linge; Breen, David (2008). "Calculating Center of Mass in an
-            # Unbounded 2D Environment". Journal of Graphics, GPU, and Game Tools. 13
-            # (4): 53–60. doi:10.1080/2151237X.2008.10129266,
-            # https://en.wikipedia.org/wiki/Center_of_mass
-            # Systems_with_periodic_boundary_conditions
-
-            if self.comgroup is not None:
-                Theta = self.comsel.positions[:,
-                                              self.dim] / self._ts.dimensions[
-                                                  self.dim] * 2 * np.pi
-                Xi = (np.cos(Theta) *
-                      self.comsel.masses).sum() / self.comsel.masses.sum()
-                Zeta = (np.sin(Theta) *
-                        self.comsel.masses).sum() / self.comsel.masses.sum()
-                ThetaCOM = np.arctan2(-Zeta, -Xi) + np.pi
-                comshift = self._ts.dimensions[self.dim] * (0.5 - ThetaCOM / (2 * np.pi))
-
-                # Check if SingleGroupAnalysis
-                if hasattr(self, 'atomgroup'):
-                    groups = [self.atomgroup]
-                else:
-                    groups = self.atomgroups
-                for group in groups:
-                    group.atoms.positions += comshift
-
-            # Call the original _single_frame
-            orig_single_frame(self, *args, **kwargs)
-
-        def _calculate_results(self):
-            self._index = self._frame_index + 1
-
-            if self._zmax is None:
-                zmax = self.Lz / self._index
-            else:
-                zmax = self.zmax
-
-            dz = (zmax - self.zmin) / self.n_bins
-
-            self.results["z"] = np.linspace(
-                self.zmin+dz/2, zmax-dz/2, self.n_bins,
-                endpoint=False)
-
-            if self.center:
-                self.results["z"] -= self.zmin + (zmax - self.zmin) / 2
-
-            self.results["z"] /= 10
-            orig_calculate_results(self)
-
-        extra_params = [
-        {
-            "name": "dim",
-            "type": int,
-            "doc": "Dimension for binning (0=X, 1=Y, 2=Z)\n"
-        }, {
-            "name": "binwidth",
-            "type": float,
-            "doc": "binwidth (nanometer)\n"
-        }, {
-            "name": "comgroup",
-            "type": str,
-            "doc": "Perform the binning relative to the "
-                   "center of mass of the selected group."
-                   "With `comgroup` the `center` option is also used.\n"
-        }, {
-            "name": "center",
-            "type": bool,
-            "doc": "Perform the binning relative to the center of the "
-                   "(changing) box.\n"
-        }]
-
-        original_class.__doc__ = _append_to_doc(original_class.__doc__,
-                                                params=extra_params)
-
-        # Set the class functions to the new ones
-        original_class.__init__ = __init__
-        original_class._configure_parser = _configure_parser
-        original_class._prepare = _prepare
-        original_class.get_bins = get_bins
-        original_class._single_frame = _single_frame
-        original_class._calculate_results = _calculate_results
-
-        return original_class
-
-    return inner
+def set_planar_class_doc(public_api):
+    if public_api.__doc__ is not None:
+        public_api.__doc__ = public_api.__doc__.replace(
+            "${PLANAR_CLASS_PARAMETERS}",
+            planar_class_parameters_doc)
+        public_api.__doc__ = public_api.__doc__.replace(
+            "${PLANAR_CLASS_ATTRIBUTES}",
+            planar_class_attributes_doc)
+    return public_api
 
 
 def charge_neutral(filter):
@@ -221,15 +69,17 @@ def charge_neutral(filter):
     with the filter attribute. If the AtomGroup's corresponding universe is non-neutral
     an ValueError is raised.
 
-    :param filter (str): Filter type to control warning filter
-                         Common values are: "error" or "default"
-                         See `warnings.simplefilter` for more options.
+    Parameters
+    ----------
+    filter : str
+        Filter type to control warning filter
+        Common values are: "error" or "default"
+        See `warnings.simplefilter` for more options.
     """
     def inner(original_class):
         def charge_check(function):
             @functools.wraps(function)
             def wrapped(self):
-                # Check if SingleGroupAnalysis
                 if hasattr(self, 'atomgroup'):
                     groups = [self.atomgroup]
                 else:
@@ -255,6 +105,34 @@ def charge_neutral(filter):
             return wrapped
 
         original_class._prepare = charge_check(original_class._prepare)
+
+        return original_class
+
+    return inner
+
+
+def make_whole():
+    """Class Decorator to make molecules whole in each analysis step."""
+    def inner(original_class):
+        def make_whole(function):
+            @functools.wraps(function)
+            def wrapped(self):
+                if hasattr(self, 'make_whole') and self.make_whole:
+                    if hasattr(self, "atomgroup"):
+                        groups = [self.atomgroup]
+                    else:
+                        groups = self.atomgroups
+                    for group in groups:
+                        group.unwrap(compound=check_compound(group))
+                return function(self)
+
+            return wrapped
+
+        if original_class.__doc__ is not None:
+            original_class.__doc__ = original_class.__doc__.replace(
+                "${MAKE_WHOLE_PARAMETER}",
+                make_whole_parameter_doc)
+        original_class._single_frame = make_whole(original_class._single_frame)
 
         return original_class
 
