@@ -10,11 +10,41 @@
 import numpy as np
 
 from ..utils import check_compound, savetxt
-from .base import SingleGroupAnalysisBase
+from ..decorators import set_verbose_doc
+from .base import AnalysisBase
 
 
-class dipole_angle(SingleGroupAnalysisBase):
+@set_verbose_doc
+class DipoleAngle(AnalysisBase):
     """Calculate angle timeseries of dipole moments with respect to an axis.
+
+    The dipole angle function in the timeseries module computes the dipole
+    moment of an MD simulation trajectory with respect to a reference axis.
+
+    Parameters
+    ----------
+    atomgroup : AtomGroup
+       Atomgroup on which the analysis is executed
+    dim : int
+        refernce vector for angle (x=0, y=1, z=2)
+    output : str
+       Prefix for output filenames
+    concfreq : int
+        Default number of frames after which results are calculated and files refreshed.
+        If `0` results are only calculated at the end of the analysis and not
+        saved by default.
+    ${VERBOSE_PARAMETER}
+
+    Attributes
+    ----------
+    results.t : np.ndarray
+        time (ps)
+    resulst.cos_theta_i : np.ndarray
+        Average cos between dipole and axis
+    resulst.cos_theta_ii : np.ndarray
+        Average cos^2 of the same between dipole and axis
+    resulst.cos_theta_ij : np.ndarray
+        Product cos of dipole i and cos of dipole j (i!=j)
 
     The dipole angle function in the timeseries module computes the dipole
     moment of a MD simulation trajectory with respect to a reference axis.
@@ -93,39 +123,17 @@ class dipole_angle(SingleGroupAnalysisBase):
     .. _`MDAnalysis`: https://www.mdanalysis.org/
     .. _`matplotlib`: https://matplotlib.org/
     .. _`gromacs electric field`: https://manual.gromacs.org/2019-current/reference-manual/special/electric-fields.html#fig-field
-
-
-    **Inputs**
-
-    :param dim (int): refernce vector for angle (x=0, y=1, z=2)
-    :param outfreq (float): Default number of frames after which output files are refreshed
-    :param output (str): Prefix for output filenames
-
-    **Attributes**
-
-    :returns results.t (numpy.ndarray): time (ps)
-    :returns results.cos_theta_i (numpy.ndarray): Average cos between dipole and axis
-    :returns results.cos_theta_ii (numpy.ndarray): Average cos^2 of the same between dipole and axis
-    :returns results.cos_theta_ij (numpy.ndarray): Product cos of dipole i and cos of dipole j (i!=j)
-
     """
-
     def __init__(self,
                  atomgroup,
-                 output="dipangle.dat",
-                 outfreq=10000,
                  dim=2,
+                 output="dipangle.dat",
+                 concfreq=0,
                  **kwargs):
-        super().__init__(atomgroup, **kwargs)
-        self.output = output
+        super(DipoleAngle, self).__init__(atomgroup, **kwargs)
         self.dim = dim
-        self.outfreq = outfreq
-
-    def _configure_parser(self, parser):
-        parser.description = self.__doc__
-        parser.add_argument('-d', dest='dim')
-        parser.add_argument('-dout', dest='outfreq')
-        parser.add_argument('-o', dest='output')
+        self.output = output
+        self.concfreq = concfreq
 
     def _prepare(self):
         self.n_residues = self.atomgroup.residues.n_residues
@@ -159,37 +167,57 @@ class dipole_angle(SingleGroupAnalysisBase):
         self.cos_theta_ij[self._frame_index] /= (self.n_residues**2 -
                                                  self.n_residues)
 
-        if self._save and self._frame_index % self.outfreq == 0 and self._frame_index > 0:
-            self._calculate_results()
-            self._save_results()
+        if self.concfreq and self._frame_index % self.concfreq == 0 \
+            and self._frame_index > 0:
+            self._conclude()
+            self.save()
 
-    def _calculate_results(self):
+    def _conclude(self):
         self._index = self._frame_index + 1
 
-        self.results.t = self._trajectory.dt * \
-            np.arange(self.startframe, self.stopframe, self.step)
-
+        self.results.t = self.times
         self.results.cos_theta_i = self.cos_theta_i[:self._index]
         self.results.cos_theta_ii = self.cos_theta_ii[:self._index]
         self.results.cos_theta_ij = self.cos_theta_ij[:self._index]
 
-    def _save_results(self):
-
+    def save(self):
         savetxt(self.output,
-                np.vstack([
-                    self.results.t, self.results.cos_theta_i,
-                    self.results.cos_theta_ii, self.results.cos_theta_ij
-                ]).T,
+                np.vstack([self.results.t,
+                           self.results.cos_theta_i,
+                           self.results.cos_theta_ii,
+                           self.results.cos_theta_ij]).T,
                 header="t\t<cos(θ_i)>\t<cos(θ_i)cos(θ_i)>\t<cos(θ_i)cos(θ_j)>",
                 fmt='%.5e')
 
 
-class kinetic_energy(SingleGroupAnalysisBase):
+class KineticEnergy(AnalysisBase):
     """Computes the timeseries of energies.
 
     The kinetic energy function computes the translational and rotational kinetic
     energy with respect to molecular center (center of mass, center of charge)
     of an MD simulation trajectory.
+
+    Parameters
+    ----------
+    atomgroup : AtomGroup
+       Atomgroup on which the analysis is executed
+    refpoint : str
+        reference point for molecular center: center of
+        mass (COM), center of charge (COC), or oxygen position (OXY)
+        Note: The oxygen position only works for systems of pure water
+    output : str
+        Output filename
+    ${VERBOSE_PARAMETER}
+
+    Attributes
+    ----------
+    results.t : np.ndarray
+        time (ps)
+    results.trans : np.ndarray
+        translational kinetic energy (kJ/mol)
+    results.rot : np.ndarray
+        rotational kinetic energy (kJ/mol)
+
 
     As an example, we analyse a box of water molecules containing around 500
     water molecules and simulated for 200 ps in the NVE ensemble
@@ -253,7 +281,8 @@ class kinetic_energy(SingleGroupAnalysisBase):
 
         maicos kinetic_energy -s md.tpr -f md.trr -o ke.dat
 
-    The output file ``ke.dat`` is similar to ``ke.results`` and contains the data in columns.
+    The output file ``ke.dat`` is similar to ``ke.results`` and contains
+    the data in columns.
 
     They are several options you can play with. To know the full
     list of options, have a look at the ``Inputs`` section.
@@ -261,31 +290,12 @@ class kinetic_energy(SingleGroupAnalysisBase):
 
     .. _`MDAnalysis`: https://www.mdanalysis.org/
     .. _`matplotlib`: https://matplotlib.org/
-
-
-    **Inputs**
-
-    :param output (str): Output filename
-    :param refpoint (str): reference point for molecular center: center of
-                            mass (COM), center of charge (COC), or oxygen position (OXY)
-                            Note: The oxygen position only works for systems of pure water
-
-    **Attributes**
-
-    :returns results.t (numpy.ndarray): time (ps)
-    :returns results.trans (numpy.ndarray): translational kinetic energy (kJ/mole)
-    :returns results.rot (numpy.ndarray): rotational kinetic energy (kJ/mole)
-
     """
 
     def __init__(self, atomgroup, output="ke.dat", refpoint="COM", **kwargs):
-        super().__init__(atomgroup, **kwargs)
+        super(KineticEnergy, self).__init__(atomgroup, **kwargs)
         self.output = output
         self.refpoint = refpoint
-
-    def _configure_parser(self, parser):
-        parser.add_argument('-o', dest='output')
-        parser.add_argument('-r', dest='refpoint')
 
     def _prepare(self):
         """Set things up before the analysis loop begins"""
@@ -333,16 +343,15 @@ class kinetic_energy(SingleGroupAnalysisBase):
                                                   np.linalg.norm(v, axis=1)**2)
 
     def _calculate_results(self):
-        self.results.t= self._trajectory.dt * \
-            np.arange(self.startframe, self.stopframe, self.step)
+        self.results.t= self.times
         self.results.trans = self.E_center / 2 / 100
         self.results.rot = (self.E_kin - self.E_center) / 2 / 100
 
-    def _save_results(self):
+    def save(self):
         savetxt(self.output,
                 np.vstack([
                     self.results.t, self.results.trans,
                     self.results.rot
                 ]).T,
                 fmt='%.8e',
-                header="t / ps \t E_kin^trans / kJ/mole \t E_kin^rot / kJ/mole")
+                header="t / ps \t E_kin^trans / kJ/mol \t E_kin^rot / kJ/mole")
