@@ -10,20 +10,31 @@
 import MDAnalysis as mda
 import pytest
 
-from maicos import dipole_angle
+from maicos import DipoleAngle, KineticEnergy
 import numpy as np
 from numpy.testing import assert_almost_equal
 
-from datafiles import WATER_GRO, WATER_TPR, WATER_TRR
+from datafiles import WATER_TPR, WATER_TRR
 
-class Test_dipole_angle(object):
+class TestDipoleAngle(object):
 
     @pytest.fixture()
     def ag(self):
         u = mda.Universe(WATER_TPR, WATER_TRR)
         return u.select_atoms('type OW or type H')
 
-    def test_dipole_angle(self, ag):
-        dipa = dipole_angle(ag)
+    def test_DipoleAngle(self, ag):
+        dipa = DipoleAngle(ag)
         dipa.run()
         assert_almost_equal(np.sum(dipa.cos_theta_i),-3.99,decimal=2)
+
+
+class TestKineticEnergy(object):
+
+    @pytest.fixture()
+    def ag(self):
+        u = mda.Universe(WATER_TPR, WATER_TRR)
+        return u.atoms
+
+    def test_KineticEnergy(self, ag):
+        KineticEnergy(ag).run()
