@@ -10,14 +10,14 @@
 import MDAnalysis as mda
 import pytest
 
-from maicos import velocity
+from maicos import Velocity
 import numpy as np
 from numpy.testing import assert_almost_equal
 
 from datafiles import WATER_TPR, WATER_TRR
 
 
-class Test_velocity(object):
+class TestVelocity(object):
 
     @pytest.fixture()
     def vel_array(self):
@@ -38,29 +38,33 @@ class Test_velocity(object):
     @pytest.mark.parametrize('dim', (0, 1, 2))
     @pytest.mark.parametrize('vdim', (0, 1, 2))
     def test_dens(self, ag, dim, vdim, vel_array):
-        vel = velocity(ag, dim=dim, vdim=vdim).run()
+        vel = Velocity(ag, dim=dim, vdim=vdim).run()
         assert_almost_equal(vel.results['v'].mean(),
                             vel_array[dim][vdim],
                             decimal=0)
 
     def test_broken_molecules(self, ag):
-        vel = velocity(ag, bpbc=False).run()
+        vel = Velocity(ag, bpbc=False).run()
         assert_almost_equal(vel.results['v'].mean(), -0.16, decimal=1)
 
     def test_repaired_molecules(self, ag):
-        vel = velocity(ag, bpbc=True).run()
+        vel = Velocity(ag, bpbc=True).run()
         assert_almost_equal(vel.results['v'].mean(), -0.26, decimal=1)
 
     def test_output(self, ag, tmpdir):
         with tmpdir.as_cwd():
-            vel = velocity(ag, end=20, save=True, mu=True).run()
+            vel = Velocity(ag, end=20, save=True, mu=True)
+            vel.run()
+            vel.save()
             res_vel = np.loadtxt("vel_{}.dat".format(vel.output_suffix))
             assert_almost_equal(vel.results["v"], res_vel[:, 1], decimal=2)
 
     def test_output_name(self, ag, tmpdir):
         with tmpdir.as_cwd():
-            velocity(ag, output_suffix="foo", end=20, save=True).run()
+            vel =Velocity(ag, output_suffix="foo", end=20, save=True).run()
+            vel.run()
+            vel.save()
             open("vel_foo.dat")
 
     def test_verbose(self, ag):
-        velocity(ag, verbose=True).run()
+        Velocity(ag, verbose=True).run()
